+++ bin/cli.php
@@ -27,6 +27,7 @@ use App\Console\MigrateCreatedDay;
 use App\Console\MigrateMailLogData;
 use App\Console\MigrateIdActionIndex;
 use App\Console\MigrateMailLogTokens;
+use App\Console\MigrateIpCreatedDayIndex;
 use App\Console\MigrateDb;
 
 (new Kernel())->boot();
@@ -45,6 +46,7 @@ $application->add(new MigrateCreatedDay());
 $application->add(new MigrateMailLogData());
 $application->add(new MigrateIdActionIndex());
 $application->add(new MigrateMailLogTokens());
+$application->add(new MigrateIpCreatedDayIndex());
 $application->add(new MigrateDb());
 
 $application->run();