+++ contrib/db-init.sql
@@ -175,7 +175,8 @@ CREATE TABLE `mail_logs` (
   KEY `created_day_action_index` (`created_day`,`action`),
   KEY `mail_stored_created_day_index` (`mail_stored`,`created_day`),
   KEY `has_virus_created_day_index` (`has_virus`,`created_day`),
-  KEY `id_action_index` (`id`,`action`)
+  KEY `id_action_index` (`id`,`action`),
+  KEY `ip_created_day_index` (`ip`,`created_day`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci ROW_FORMAT=DYNAMIC;
 
 CREATE TABLE `mail_log_recipients` (
@@ -230,4 +231,5 @@ INSERT INTO `migrations` VALUES
 ('20260729_created_day','completed',NOW()),
 ('20260731_mail_log_data','completed',NOW()),
 ('20260806_id_action_index','completed',NOW()),
-('20260904_mail_log_tokens','completed',NOW());
+('20260904_mail_log_tokens','completed',NOW()),
+('20260906_ip_created_day_index','completed',NOW());